Fix: Fatal error with item id saved on Core Data
Int16 wasn't enough, then the error "Not enough bits to represent the passed value" was given, an Int64 is now used.

+++ CurrentNewsFeed/ViewControllers/ItemViewController.swift
@@ -98,7 +98,7 @@ final class ItemViewController: UIViewController {
     
     private func save() {
         let favorite = Favorite(context: DataManager.context)
-        favorite.itemID = Int16(self.item.id)
+        favorite.itemID = Int64(self.item.id)
         favorite.title = self.item.title
         favorite.type = self.item.type
         favorite.url = self.item.url
